Adds working tag to htk_tags.
